Refactor everything to modules
